MedVault-main/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.1.0
bitarray==3.0.0
cbor2==5.6.5
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
crispy-bootstrap4==2024.10
cytoolz==1.0.1
Django==5.1.5
django-crispy-forms==2.3
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
hexbytes==1.3.0
idna==3.10
importlib_metadata==8.6.1
multidict==6.1.0
packaging==23.2
parsimonious==0.10.0
pillow==11.1.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
pyunormalize==16.0.0
pyzbar==0.1.9
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
vyper==0.3.9
web3==7.7.0
websockets==13.1
wheel==0.45.1
yarl==1.18.3
zipp==3.21.0
